requirements.txt
pillow
fastapi
opencv-python
tensorflow==2.10
scipy
scikit-image
onnxruntime-gpu
python-multipart
uvicorn
torch